% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_ess}
\alias{calculate_ess}
\title{Calculate Effective Sample Size}
\usage{
calculate_ess(chains)
}
\arguments{
\item{chains}{List of MCMC chains}
}
\value{
Effective Sample Size
}
\description{
Calculate Effective Sample Size
}
\keyword{internal}
