% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional_diagnostics.R,
%   R/enhanced_trace_plot.R
\name{enhanced_trace_plot}
\alias{enhanced_trace_plot}
\title{Enhanced Trace Plot for Multiple Chains}
\usage{
enhanced_trace_plot(chains, param_names)

enhanced_trace_plot(chains, param_names)
}
\arguments{
\item{chains}{A list of MCMC chains}

\item{param_names}{A vector of parameter names}

\item{parameter_names}{A vector of parameter names}
}
\value{
A ggplot object

A ggplot object
}
\description{
Enhanced Trace Plot for Multiple Chains

Enhanced Trace Plot
}
