% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_examples.R
\name{generate_mcmc_chain}
\alias{generate_mcmc_chain}
\title{Generate a simple MCMC chain using Metropolis-Hastings}
\usage{
generate_mcmc_chain(n, a, b, start, proposal_sd)
}
\arguments{
\item{n}{Number of iterations}

\item{a}{Shape parameter for Beta distribution}

\item{b}{Shape parameter for Beta distribution}

\item{start}{Starting value}

\item{proposal_sd}{Standard deviation for proposal distribution}
}
\value{
A list containing the chain and acceptance indicators
}
\description{
Generate a simple MCMC chain using Metropolis-Hastings
}
