% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional_diagnostics.R
\name{geweke_diagnostic}
\alias{geweke_diagnostic}
\title{Geweke Diagnostic}
\usage{
geweke_diagnostic(chain, first = 0.1, last = 0.5)
}
\arguments{
\item{chain}{A numeric vector representing an MCMC chain}

\item{first}{Proportion of chain to use for first segment (default 0.1)}

\item{last}{Proportion of chain to use for last segment (default 0.5)}
}
\value{
A numeric value representing the Geweke statistic
}
\description{
Geweke Diagnostic
}
