% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional_diagnostics.R
\name{heidelberger_welch_test}
\alias{heidelberger_welch_test}
\title{Heidelberger and Welch Test}
\usage{
heidelberger_welch_test(chain, alpha = 0.05)
}
\arguments{
\item{chain}{A numeric vector representing an MCMC chain}

\item{alpha}{Significance level for the test (default 0.05)}
}
\value{
A list containing test results
}
\description{
Heidelberger and Welch Test
}
