% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_diag.R
\name{mcmc_diag}
\alias{mcmc_diag}
\title{MCMC Diagnostic Object}
\usage{
mcmc_diag(chains)
}
\arguments{
\item{chains}{List of MCMC chains}
}
\value{
An object of class "mcmc_diag"
}
\description{
MCMC Diagnostic Object
}
