\name{mcmcdiag-package}
\alias{mcmcdiag-package}
\alias{mcmcdiag}
\docType{package}
\title{
\packageTitle{mcmcdiag}
}
\description{Uses batch means to calculate a convergence diagnostic for Markov chain Monte Carlo in the spirit of Gelman-Rubin
\packageDescription{mcmcdiag}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{mcmcdiag}
\packageIndices{mcmcdiag}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{mcmcdiag}

Maintainer: \packageMaintainer{mcmcdiag}
}
\references{ 
  Flegal, J. M. and Jones, G. L. (2010) Batch means and
  spectral variance estimators in Markov chain Monte Carlo.
  \emph{The Annals of Statistics}, \bold{38}, 1034--1070.

Gelman, A and Rubin, DB (1992) Inference from iterative simulation
using multiple sequences, \emph{Statistical Science}, \bold{7}, 457-511.

Brooks, SP. and Gelman, A. (1998) General methods for monitoring
convergence of iterative simulations. \emph{Journal of Computational and
Graphical Statistics}, \bold{7}, 434-455.
}


\keyword{ htest }
\examples{

}
