% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_diag.R
\name{plot.mcmc_diag}
\alias{plot.mcmc_diag}
\title{Plot method for mcmc_diag objects}
\usage{
\method{plot}{mcmc_diag}(x, ...)
}
\arguments{
\item{x}{An object of class "mcmc_diag"}

\item{...}{Further arguments passed to or from other methods}
}
\description{
Plot method for mcmc_diag objects
}
