% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_diag.R
\name{print.mcmc_diag}
\alias{print.mcmc_diag}
\title{Print method for mcmc_diag objects}
\usage{
\method{print}{mcmc_diag}(x, ...)
}
\arguments{
\item{x}{An object of class "mcmc_diag"}

\item{...}{Further arguments passed to or from other methods}
}
\description{
Print method for mcmc_diag objects
}
