% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_diag.R
\name{summary.mcmc_diag}
\alias{summary.mcmc_diag}
\title{Summary method for mcmc_diag objects}
\usage{
\method{summary}{mcmc_diag}(object, ...)
}
\arguments{
\item{object}{An object of class "mcmc_diag"}

\item{...}{Further arguments passed to or from other methods}
}
\description{
Summary method for mcmc_diag objects
}
