% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target.psrf.R
\name{target.psrf}
\alias{target.psrf}
\title{Calculates a Gelman Rubin diagnostic threshold using lugsail variance estimators.}
\usage{
target.psrf(p, epsilon = 0.05, delta = NULL, m, alpha = 0.05)
}
\arguments{
\item{p}{dimension of the estimation problem.}

\item{epsilon}{relative precision level.}

\item{delta}{desired delta value.}

\item{m}{number of chains.}

\item{alpha}{significance level.}
}
\value{
\item{psrf}{The psrf threshold to be reached for convergence.}

\item{epsilon}{The epsilon value used to calculate the psrf threshold.}

\item{delta}{The delta value used to calculate the psrf threshold.}
}
\description{
When the sample diagnostic reaches the psrf threshold calculated in this function, the chain has converged.
}
\section{References}{

Vats, D. and Knudson, C. Revisiting the Gelman-Rubin Diagnostic.	arXiv:1812.09384 

Vats, D. and Flegal, J. Lugsail lag windows and their application to MCMC. arXiv: 1809.04541.

Flegal, J. M. and Jones, G. L. (2010) Batch means and spectral variance estimators in Markov chain Monte Carlo. \emph{The Annals of Statistics}, \bold{38}, 1034--1070. \cr
Gelman, A and Rubin, DB (1992) Inference from iterative simulation using multiple sequences, \emph{Statistical Science}, \bold{7}, 457-511. \cr
Brooks, SP. and Gelman, A. (1998) General methods for monitoring convergence of iterative simulations. \emph{Journal of Computational and Graphical Statistics}, \bold{7}, 434-455.
}

\examples{
target.psrf(1, delta = .1, m = 5, alpha = .05)
target.psrf(1, epsilon = .05, m = 5, alpha = .05)

}
