% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_ess}
\alias{calculate_ess}
\title{Calculate Effective Sample Size (ESS)}
\usage{
calculate_ess(chains)
}
\arguments{
\item{chains}{List of MCMC chains (numeric vectors).}
}
\value{
Effective Sample Size for the provided chains.
}
\description{
This function calculates the Effective Sample Size for MCMC chains.
}
\keyword{internal}
