% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choosepsrf.R
\name{choosepsrf}
\alias{choosepsrf}
\title{Calculates a Gelman Rubin diagnostic threshold.}
\usage{
choosepsrf(p, epsilon, m, alpha = 0.05)
}
\arguments{
\item{p}{dimension of the estimation problem}

\item{epsilon}{relative precision level}

\item{m}{number of chains}

\item{alpha}{significance level}
}
\value{
The psrf threshold to be reached for convergence.
}
\description{
When the sample diagnostic reaches this threshold, the chain has converged.
}
