% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional_diagnostics.R
\name{enhanced_trace_plot}
\alias{enhanced_trace_plot}
\title{Enhanced Trace Plot for Multiple Chains}
\usage{
enhanced_trace_plot(chains, parameter_names)
}
\arguments{
\item{chains}{A list of numeric vectors representing MCMC chains}

\item{parameter_names}{A vector of parameter names}
}
\value{
A ggplot object
}
\description{
Enhanced Trace Plot for Multiple Chains
}
