% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gelman_rubin.R, R/utils.R
\name{gelman_rubin}
\alias{gelman_rubin}
\title{Gelman-Rubin Diagnostic}
\usage{
gelman_rubin(chains)

gelman_rubin(chains)
}
\arguments{
\item{chains}{A list of MCMC chains}
}
\value{
A numeric value representing the Gelman-Rubin statistic

The Gelman-Rubin statistic
}
\description{
Gelman-Rubin Diagnostic

Gelman-Rubin Diagnostic
}
