% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gelman_rubin.R
\name{gelman_rubin}
\alias{gelman_rubin}
\title{Gelman-Rubin Diagnostic (C++ Implementation)}
\usage{
gelman_rubin(chains, burn_in = 0)
}
\arguments{
\item{chains}{A list of MCMC chains. Each chain should be a numeric vector.}
}
\value{
A numeric value representing the potential scale reduction factor (PSRF).
}
\description{
This function calculates the Gelman-Rubin diagnostic for MCMC convergence using a C++ implementation.
}
\examples{
chains <- list(rnorm(1000), rnorm(1000), rnorm(1000))
gelman_rubin_cpp(chains)
}
