% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_diag.R
\name{mcmc_diag}
\alias{mcmc_diag}
\title{Comprehensive MCMC Diagnostics}
\usage{
mcmc_diag(chains)
}
\arguments{
\item{chains}{A list of MCMC chains}
}
\value{
A list containing diagnostic results
}
\description{
Comprehensive MCMC Diagnostics
}
