\name{mcmcdiag-package}
\alias{mcmcdiag-package}
\alias{mcmcdiag}
\docType{package}
\title{
\packageTitle{mcmcdiag}
}
\description{
\packageDescription{mcmcdiag}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{mcmcdiag}
\packageIndices{mcmcdiag}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{mcmcdiag}

Maintainer: \packageMaintainer{mcmcdiag}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
