\name{mcmcdiag-package}
\alias{mcmcdiag-package}
\alias{mcmcdiag}
\docType{package}
\title{
\packageTitle{mcmcdiag}
}
\description{Calculates a convergence diagnostic in the spirit of Gelman-Rubin
\packageDescription{mcmcdiag}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{mcmcdiag}
\packageIndices{mcmcdiag}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{mcmcdiag}

Maintainer: \packageMaintainer{mcmcdiag}
}
\references{
Gelman Rubin 1992; Brooks and Gelman 1998 (add these later)
}

\keyword{ htest }
\examples{

}
