% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_plot.R
\name{trace_plot}
\alias{trace_plot}
\title{Generate Trace Plot for MCMC Chain}
\usage{
trace_plot(chain, param_name)
}
\arguments{
\item{chain}{Numeric vector representing MCMC chain values.}

\item{parameter}{Name of the parameter being plotted.}
}
\value{
A trace plot using base R graphics.
}
\description{
This function generates a trace plot for a given MCMC chain.
}
