% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_plot.R
\name{trace_plot}
\alias{trace_plot}
\title{Trace Plot}
\usage{
trace_plot(chain, parameter)
}
\arguments{
\item{chain}{MCMC chain}

\item{parameter}{Name of the parameter to plot}
}
\value{
ggplot object
}
\description{
Trace Plot
}
